+++ pkg/kt_errors/fault_builder.go
@@ -11,7 +11,7 @@ import (
 // You must specify the kind of the error right away - this is not changeable later.
 func NewPublicFaultBuilder(errType FaultKind) *FaultBuilder {
 	err := newInitializedFault(errType)
-	err.isPublic = true
+	err.public = true
 	return &FaultBuilder{fault: err, errCodes: ktsets.NewSet[string]()}
 }
 
@@ -19,7 +19,7 @@ func NewPublicFaultBuilder(errType FaultKind) *FaultBuilder {
 // You must specify the kind of the error right away - this is not changeable later.
 func NewFaultBuilder(errType FaultKind) *FaultBuilder {
 	err := newInitializedFault(errType)
-	err.isPublic = false
+	err.public = false
 	return &FaultBuilder{fault: err, errCodes: ktsets.NewSet[string]()}
 }
 
@@ -29,17 +29,17 @@ type FaultBuilder struct {
 }
 
 func (builder *FaultBuilder) Build() Fault {
-	builder.fault.errorCodes = builder.errCodes.GetAll()
+	builder.fault.ErrorCodes = builder.errCodes.GetAll()
 	// review the isRetryable flag
-	if builder.fault.isRetryable {
-		switch builder.fault.kind {
+	if builder.fault.Retryable {
+		switch builder.fault.Kind {
 		case AuthenticationFault:
 			if builder.errCodes.ContainsAny(AUTHENTICATION_ERRCODE_MISSING, AUTHENTICATION_ERRCODE_NOT_SUPPORTED) {
-				builder.fault.isRetryable = false
+				builder.fault.Retryable = false
 			}
 		case AuthorizationFault:
 			if builder.errCodes.ContainsAny(AUTHORIZATION_NO_PERMISSION) {
-				builder.fault.isRetryable = false
+				builder.fault.Retryable = false
 			}
 		}
 	}
@@ -51,12 +51,12 @@ func (builder *FaultBuilder) Build() Fault {
 // Please note: certain error types are inheritedly not retryable, e.g. ValidationError or NotImplementedError. Invoking this method
 // on any of those will simply have no effect.
 func (builder *FaultBuilder) WithIsRetryable(flag bool) *FaultBuilder {
-	switch builder.fault.kind {
+	switch builder.fault.Kind {
 	// only these types can be classified as retryable
 	case NotImplementedFault, ValidationFault, ResourceNotFoundFault:
 		// we skip it - these are inheritedly not retryable
 	default:
-		builder.fault.isRetryable = flag
+		builder.fault.Retryable = flag
 	}
 	return builder
 }
@@ -65,34 +65,34 @@ func (builder *FaultBuilder) WithIsRetryable(flag bool) *FaultBuilder {
 // Why is it a "template"? Because you can use variables in it (Python style), e.g. "My string with {var1} and {var2} variables.".
 // Then add these as labels (see `WithLabel()` / `WithLabels()`).
 func (builder *FaultBuilder) WithMessageTemplate(msg string) *FaultBuilder {
-	builder.fault.messageTemplate = msg
+	builder.fault.MessageTemplate = msg
 	return builder
 }
 
 // Sets a message template for a specific audience.
 func (builder *FaultBuilder) WithMessageTemplateForAudience(forAudience string, msg string) *FaultBuilder {
-	builder.fault.messageTemplatesByAudience[forAudience] = msg
+	builder.fault.MessageTemplatesByAudience[forAudience] = msg
 	return builder
 }
 
 // If you changed your mind you can remove the template for this audience
 func (builder *FaultBuilder) WithoutMessageTemplateForAudiences(forAudiences ...string) *FaultBuilder {
 	for _, audience := range forAudiences {
-		delete(builder.fault.messageTemplatesByAudience, audience)
+		delete(builder.fault.MessageTemplatesByAudience, audience)
 	}
 	return builder
 }
 
 // Adds all audience message templates to the error - this is a merge.
 func (builder *FaultBuilder) WithMessageTemplatesByAudience(templates map[string]string) *FaultBuilder {
-	maps.Copy(builder.fault.messageTemplatesByAudience, templates)
+	maps.Copy(builder.fault.MessageTemplatesByAudience, templates)
 	return builder
 }
 
 // Adds all audience message templates to the error - and these will override the possibly existing ones.
 func (builder *FaultBuilder) WithExactMessageTemplatesByAudience(templates map[string]string) *FaultBuilder {
-	builder.fault.messageTemplatesByAudience = make(map[string]string, len(templates))
-	maps.Copy(builder.fault.messageTemplatesByAudience, templates)
+	builder.fault.MessageTemplatesByAudience = make(map[string]string, len(templates))
+	maps.Copy(builder.fault.MessageTemplatesByAudience, templates)
 	return builder
 }
 
@@ -134,14 +134,14 @@ func (builder *FaultBuilder) WithoutErrorCodes(c ...string) *FaultBuilder {
 
 // Attaching a label (key-value pair) to this error.
 func (builder *FaultBuilder) WithLabel(key string, value any) *FaultBuilder {
-	builder.fault.labels[key] = value
+	builder.fault.Labels[key] = value
 	return builder
 }
 
 // If you changed your mind you can remove specific labels (key-value pair) from this error.
 func (builder *FaultBuilder) WithoutLabels(keys ...string) *FaultBuilder {
 	for _, key := range keys {
-		delete(builder.fault.labels, key)
+		delete(builder.fault.Labels, key)
 	}
 	return builder
 }
@@ -149,13 +149,13 @@ func (builder *FaultBuilder) WithoutLabels(keys ...string) *FaultBuilder {
 // You can attach multiple labels (key-value pairs) in one go if you wish with this method.
 // Please note that these will be simply merged into the existing labels! See also `WithExactLabels()` method!
 func (builder *FaultBuilder) WithLabels(labels map[string]any) *FaultBuilder {
-	maps.Copy(builder.fault.labels, labels)
+	maps.Copy(builder.fault.Labels, labels)
 	return builder
 }
 
 // Sets the labels (key-value pairs) attached to this error to the given map - all previous labels will be removed.
 func (builder *FaultBuilder) WithExactLabels(labels map[string]any) *FaultBuilder {
-	builder.fault.labels = make(map[string]any, len(labels))
-	maps.Copy(builder.fault.labels, labels)
+	builder.fault.Labels = make(map[string]any, len(labels))
+	maps.Copy(builder.fault.Labels, labels)
 	return builder
 }